requirements.txt
websockets
fastapi
uvicorn
ffmpeg-python
yt-dlp
